[update] 重写 Help 模块
[更新] 重写 Help 模块

重写 Help 模块以提供更好的帮助信息

这次提交重新编写了 Help 模块，以提供更清晰、更详细的帮助信息。更新后的模块能够显示命令列表、命令说明和用法，并添加了额外的许可协议和版权声明信息。此更新可以帮助用户更好地使用 XDbot2，并增强用户体验。
