+++ src/plugins/Core/lang/zh_hans.json
@@ -35,11 +35,21 @@
   "forward._user": "用户：",
   "forward.group": "「群消息转发」",
   "forward._group": "群聊：",
-  "help.name": "命令列表",
-  "help.command": "使用 /help <命令> 获取更多信息",
-  "help.info": "说明：{}",
-  "help.usage": "用法（{}）：\n{}",
-  "help.unknown": "未知指令：{}",
+  
+  "_help.name": "命令列表",
+  "_help.command": "使用 /help <命令> 获取更多信息",
+  "_help.info": "说明：{}",
+  "_help.usage": "用法（{}）：\n{}",
+  "_help.unknown": "未知指令：{}",
+
+  "help.list": "命令列表 —— XDbot2\n{}\n使用 {}help <命令> 获取更多信息",
+  "help.list_item": "[{}] {}: {}",
+  "help.command_status": "「状态说明」\n[√]: 正常工作\n[X]: 无法使用\n[O]: 受限",
+  "help.eula": "「许可协议」\n开始使用 XDbot2 即代表您已阅读并同意《XDbot2 最终许可协议》\nhttps://github.com/ITCraftDevelopmentTeam/XDbot2/discussions/501",
+  "help.copyright": "「版权声明」\nXDbot2 由 IT Craft Development Team (https://itcdt.top) 开发，基于 MIT 许可证开源",
+  "help.info": "[{}] {}: {}\n\n用法: \n{}",
+  "help.unknown_command": "未知指令: {}",
+
   "jrrp.notice": "开始使用「今日人品」即代表您已阅读并同意《XDbot2「今日人品」用户协议》（https://github.com/ITCraftDevelopmentTeam/XDbot2/discussions/468）。祝您玩得开心！",
   "jrrp.new_record": "个人最高记录已刷新",
   "jrrp.num.100": "{}！100！！100！！！",
